requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-barcode==0.15.1
Pillow==10.1.0
openpyxl==3.1.2
Werkzeug==2.3.7
